Add optional selinux support
Since we write the target file to a new temporary file that we then
replace with the original file with an atomic move, we not only want to
preserve owner, group and mode, we also want to preserve the SELinux
context, if the operatingsystem uses selinux.

We now do this by getting the selinux context of the original file and
then apply it to the temporary file as well.